[refactor] : changed file structure
